Relocate mocker assertions allowing to prevent them being exposed.
Rather than exporting the assertions as part of unexpected-mitm put
them in a separate file that we only attach to the internal expect.
