+++ lib/unexpectedMitm.js
@@ -9,7 +9,6 @@ var messy = require('messy'),
     formatHeaderObj = require('./formatHeaderObj'),
     fs = require('fs'),
     path = require('path'),
-    trimHeadersLower = require('./trimHeadersLower'),
     memoizeSync = require('memoizesync'),
     callsite = require('callsite'),
     detectIndent = require('detect-indent'),
@@ -198,7 +197,8 @@ module.exports = {
         var expectForRendering = expect.child();
 
         expect = expect.child()
-            .use(require('unexpected-messy'));
+            .use(require('unexpected-messy'))
+            .use(require('./mockerAssertions'));
 
         expectForRendering.addType({
             name: 'infiniteBuffer',
@@ -631,80 +631,7 @@ module.exports = {
                 }
 
                 return assertionPromise;
-            })
-            .exportType({
-                name: 'UnexpectedMitmMocker',
-                base: 'object',
-                identify: function (obj) {
-                    return obj instanceof UnexpectedMitmMocker;
-                }
-            })
-            .exportAssertion('<UnexpectedMitmMocker> to be complete [with extra info]', function (expect, subject) {
-                var shouldReturnExtraInfo = expect.flags['with extra info'];
-
-                return expect.promise(function () {
-                    return subject;
-                }).then(function (result) {
-                    return [result.timeline, result.fulfilmentValue];
-                }).spread(function (timeline, fulfilmentValue) {
-                    var lastEventOrError = null;
-
-                    // pull out the last event if it exists
-                    if (timeline.length > 0) {
-                        lastEventOrError = timeline[timeline.length - 1];
-                    }
-
-                    if (lastEventOrError instanceof Error) {
-                        var name = lastEventOrError.name;
-                        if (name === 'Error' || name === 'UnexpectedError') {
-                            throw lastEventOrError;
-                        } else if (name === 'EarlyExitError') {
-                            // in the case of an early exirt error we need
-                            // to generate a diff from the last recorded
-                            // event & spec
-                            var failedEvent = timeline[timeline.length - 2];
-                            var failedExchange = failedEvent.exchange;
-                            trimHeadersLower(failedExchange.request);
-
-                            expect.errorMode = 'default';
-                            return expect(failedExchange, 'to satisfy', failedEvent.spec);
-                        } else {
-                            // ignore to cause generation of a diff
-                        }
-                    } else if (lastEventOrError === null && fulfilmentValue) {
-                        return [null, fulfilmentValue];
-                    }
-
-                    return [timeline, fulfilmentValue];
-                }).spread(function (timeline, fulfilmentValue) {
-                    // in the absence of a timeline immedistely resolve with fulfilmentValue
-                    if (timeline === null) {
-                        return fulfilmentValue;
-                    }
-
-                    var httpConversation = new messy.HttpConversation();
-                    var httpConversationSatisfySpec = { exchanges: [] };
-
-                    function recordEventForAssertion(event) {
-                        if (event.exchange) {
-                            httpConversation.exchanges.push(event.exchange);
-                        }
-                        if (event.spec) {
-                            httpConversationSatisfySpec.exchanges.push(event.spec);
-                        }
-                    }
-
-                    timeline.forEach(recordEventForAssertion);
-
-                    expect.errorMode = 'default';
-                    return expect(httpConversation, 'to satisfy', httpConversationSatisfySpec).then(function () {
-                        if (shouldReturnExtraInfo) {
-                            return [fulfilmentValue, httpConversation, httpConversationSatisfySpec];
-                        } else {
-                            return fulfilmentValue;
-                        }
-                    });
-                });
             });
+
     }
 };
